fix: reorg app into a folder so it can be packages more easily
