+++ backend/osgeo_app/blueprints/overview_map.py
