+++ backend/osgeo_app/main.py
@@ -1,0 +1,9 @@
+from . import app
+
+
+if __name__ == '__main__':
+    app = app.create_app()
+    app.run(debug=True)
+else:
+    gunicorn_app = app.create_app()
+    gunicorn_app.config['DEBUG'] = True
