+++ backend/osgeo_app/templates/plugins/leaflet.filelayer.js
