+++ backend/osgeo_app/templates/static/intersect.png
