+++ backend/osgeo_app/templates/static/layers.json
