+++ backend/osgeo_app/templates/static/lup_intersect.html
