+++ backend/osgeo_app/templates/static/modified_checkbox_example.html
