+++ backend/osgeo_app/templates/static/spiral.png
