+++ backend/osgeo_app/templates/static/style.css
